Rewrite isRiderDown() function
Completely rewrote the isRiderDown() function to 1) Work with any number of riders and 2) Be more efficient in line numbers.  Volume of crowd is now calculated in a formula based on the base volume factor, number of riders down, number of people that went down at the same time and timing gate, and the average position of all the people who crashed.