Rework Awards/Add Invalid Laps
Reworked the Awards system to be more readable.

Changed the way all player laps are added so there's no repeats in a demo so it doesn't mess up consistency calculation.

Added invalidated laps to display in qualifying at the end of the session.

Moved updating the every_rider_finished variable to a new function update_rider_finish_flags() that will update outside of a main event so we can use it for displaying invalid laps in qualifying.

Updated pro-taper bales section to sim-clips-racing bales.