Award Issues Fixed
I had some issues with awards and spectators.  It would run into some sort of error during the processing of awards, and since the displayed awards boolean doesn't happen at the end it would fail execution and keep repeating the failed execution infinitely.

Firstly I moved the displayed awards boolean to the top of the processing.  In the event that it does fail it will only happen once.  I've added a try catch statement in the rider awards so if there is an issue I can diagnose it.

I'm still not sure if it's fixed or not, but I found some probable issues with the calculations of positions gained and consistency that might've caused it, I reworked it.

The necessity for storing the slot in the rider positions after lap 1 and rider finish positions array was unnecessary so I removed it.